Relocate types pre PR feedback from @christian-bromann
