+++ packages/wdio-cli/src/interface.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk'
 import { EventEmitter } from 'events'
 import logger from '@wdio/logger'
-import type { Options, Capabilities } from '@wdio/types'
+import type { Options, Capabilities, Job } from '@wdio/types'
 
 import { getRunnerName } from './utils'
 
@@ -32,7 +32,7 @@ export default class WDIOCLInterface extends EventEmitter {
         failed: 0
     }
 
-    private _jobs: Map<string, Capabilities.Job> = new Map()
+    private _jobs: Map<string, Job> = new Map()
     private _specFileRetries: number
     private _specFileRetriesDelay: number
 
@@ -109,24 +109,24 @@ export default class WDIOCLInterface extends EventEmitter {
         this.onJobComplete(rid, this._jobs.get(rid), 0, chalk.bold.cyan('RUNNING'))
     }
 
-    onSpecRetry (rid: string, job: Capabilities.Job, retries: number) {
+    onSpecRetry (rid: string, job: Job, retries: number) {
         const delayMsg = this._specFileRetriesDelay > 0 ? ` after ${this._specFileRetriesDelay}s` : ''
         this.onJobComplete(rid, job, retries, chalk.bold(chalk.yellow('RETRYING') + delayMsg))
     }
 
-    onSpecPass (rid: string, job: Capabilities.Job, retries: number) {
+    onSpecPass (rid: string, job: Job, retries: number) {
         this.onJobComplete(rid, job, retries, chalk.bold.green('PASSED'))
     }
 
-    onSpecFailure (rid: string, job: Capabilities.Job, retries: number) {
+    onSpecFailure (rid: string, job: Job, retries: number) {
         this.onJobComplete(rid, job, retries, chalk.bold.red('FAILED'))
     }
 
-    onSpecSkip (rid: string, job: Capabilities.Job) {
+    onSpecSkip (rid: string, job: Job) {
         this.onJobComplete(rid, job, 0, 'SKIPPED', log.info)
     }
 
-    onJobComplete(cid: string, job?: Capabilities.Job, retries = 0, message = '', _logger: Function = this.log) {
+    onJobComplete(cid: string, job?: Job, retries = 0, message = '', _logger: Function = this.log) {
         const details = [`[${cid}]`, message]
         if (job) {
             details.push('in', getRunnerName(job.caps as Capabilities.DesiredCapabilities), this.getFilenames(job.specs))
@@ -158,7 +158,7 @@ export default class WDIOCLInterface extends EventEmitter {
     /**
      * add job to interface
      */
-    addJob ({ cid, caps, specs, hasTests }: Capabilities.Job & { cid: string }) {
+    addJob ({ cid, caps, specs, hasTests }: Job & { cid: string }) {
         this._jobs.set(cid, { caps, specs, hasTests })
         if (hasTests) {
             this.onSpecRunning(cid)
