+++ packages/wdio-local-runner/src/worker.ts
@@ -3,7 +3,7 @@ import child from 'child_process'
 import { EventEmitter } from 'events'
 import type { WritableStreamBuffer } from 'stream-buffers'
 import type { ChildProcess } from 'child_process'
-import type { Capabilities, Options } from '@wdio/types'
+import type { Capabilities, Options, WorkerMessageArgs, WorkerRunPayload } from '@wdio/types'
 
 import logger from '@wdio/logger'
 
@@ -58,7 +58,7 @@ export default class WorkerInstance extends EventEmitter implements Capabilities
      */
     constructor(
         config: Options.Testrunner,
-        { cid, configFile, caps, specs, execArgv, retries }: Capabilities.WorkerRunPayload,
+        { cid, configFile, caps, specs, execArgv, retries }: WorkerRunPayload,
         stdout: WritableStreamBuffer,
         stderr: WritableStreamBuffer
     ) {
@@ -184,7 +184,7 @@ export default class WorkerInstance extends EventEmitter implements Capabilities
      * @param  command  method to run in wdio-runner
      * @param  args     arguments for functions to call
      */
-    postMessage (command: string, args: Capabilities.WorkerMessageArgs): void {
+    postMessage (command: string, args: WorkerMessageArgs): void {
         const { cid, configFile, caps, specs, retries, isBusy } = this
 
         if (isBusy && command !== 'endSession') {
