+++ packages/wdio-types/src/index.ts
@@ -16,6 +16,22 @@ export type FunctionPropertyNames<T> = { [K in keyof T]: T[K] extends Function ?
 export type FunctionProperties<T> = Pick<T, FunctionPropertyNames<T>>
 export type ThenArg<T> = T extends PromiseLike<infer U> ? U : T
 
+export interface Job {
+    caps: Capabilities.DesiredCapabilities | Capabilities.W3CCapabilities | Capabilities.MultiRemoteCapabilities
+    specs: string[],
+    hasTests: boolean
+}
+
+export type WorkerMessageArgs = Omit<Job, 'caps' | 'specs' | 'hasTests'>;
+
+export interface WorkerRunPayload {
+    cid: string;
+    configFile: string;
+    caps: Capabilities.RemoteCapability;
+    specs: string[];
+    execArgv: string[];
+    retries: number;
+}
 declare global {
     namespace WebdriverIO {
         interface MochaOpts { [key: string]: any }
